Improve MPPE rejection logic.
When doing MPPE, if the peer doesn't agree to it, we need to terminate
LCP.  Older win* clients request Stac/LZS along with MPPE (because
MPPE overloads MPPC ... sigh).  So if sending CONFREJ, we would keep
LCP up if we saw a CI_MPPE *at all*, because the CONFREJ may may have
been due to the Stac/LZS option.  Now, we only keep LCP up if the MPPE
offer is acceptable.

Thanks to James Cameron for showing this problem in action.