New CHAP implementation, rewritten from scratch to avoid the code copyrighted by the uncontactable Gregory Christy.  The new code is much cleaner and splits out all the digest-specific code to separate files.  Thus the CHAP-MD5 stuff is now in chap-md5.c and all the CHAP-MS and CHAP-MSv2 stuff has moved into chap_ms.c, instead of having half of it in chap.c.
There are a few minor differences in this implementation; we don't
retransmit responses, but instead just wait for a new challenge.
The success/failure messages are more boring as well.  In fact the
digest code now sets the success/failure message.

CHAP_DIGEST_MD5 has been renamed to CHAP_MD5 for consistency.
There is a new function random_bytes() in magic.c, which generates
a string of random bytes.