Multilink improvements.  This involved moving some logic from the main loop in main.c into link_required() and link_terminated() in auth.c and adding code to multilink.c.  We now make a tdb entry with the list of pppd pids for all the links in the bundle, and the master pppd uses this to send a SIGHUP to each one when the bundle is terminated.
We still have one pppd controlling both the bundle and the first link,
but when that link goes down, assuming that other links still exist,
the first link's pppd will clean up after that link but then stay
running until all the links have disconnected.  So it is possible to
lose the first link without losing the bundle.

This requires a small kernel patch which I will be sending to the
kernel maintainers shortly.