Allow operation without an IP address for the peer
Under Linux, a point-to-point interface can operate without having
a destination IP address assigned to it, because routes can be
directed to the device rather than to a gateway IP address.  Some
peers expect us to operate in this manner and refuse to give us
an IP address for them, so this adds a new `noremoteip' option that
tells pppd not to ask for the peer's IP address if the peer doesn't
supply it.  There is also a `nosendip' option which tells pppd not
to supply its IP address -- mostly intended for testing.

Solaris requires a destination IP address, so the noremoteip option
is not included on Solaris.