Clear hungup in start_link() instead of connect_tty()
Since hungup gets set in get_input(), which is generic, but only
cleared in connect_tty(), which is specific to the tty channel type,
other channel types were showing various problems on the second and
subsequent connections when the persist option was used.  This fixes
it by clearing hungup in start_link() rather than connect_tty().