Fix authentication on second time around with multilink and persist
Frédéric Brière reported that when using multilink and persist, pppd
would fail to authenticate itself on the second and subsequent times
of bringing up a bundle.  The reason was that the doing_multilink and
multilink_master flags never got reset when the first bundle was
destroyed.  This adds lines to mp_bundle_terminated to do that.

Signed-off-by: Paul Mackerras <paulus@samba.org>