Accept ms-wins settings provided by server
The PPP servers of some UMTS/HSPA modems, including the Huawei E220
and E620, will send WINS server settings and insist that the client
accepts these. If the client does not do so, the modem will sometimes
provide bogus DNS server addresses like 10.11.12.13 and 10.11.12.14.

If we receive ms-wins settings from the server, save and include them
in our ConfReqs.

See also discussions at

  http://thread.gmane.org/gmane.linux.ppp/2721
  http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=445711

[paulus@samba.org: In fact what seems to happen is that the modem naks
IPCP conf-reqs with bogus values until it manages to communicate on the
network.  It's not clear that asking for the ms-wins stuff is essential
but it seems that sometimes it helps.]

Signed-off-by: Marcus Better <marcus@better.se>
Signed-off-by: Paul Mackerras <paulus@samba.org>