Increase default IPCP Conf-Nak limit
Some 3G modems use IPCP Conf-Naks with the same values as we requested
as a way to delay the negotiation until they have contacted a server
over their radio network.  This increases the default value for the
ipcp-max-failure option from 5 to 100 so that we don't give up too
easily when this is happening.

Signed-off-by: Paul Mackerras <paulus@samba.org>