PPPD: Fix crash when printing pppol2tp option
PPPD crashes (SEGV) when the dump or dryrun options are specified and an option
is internally defined as "o_special" with an option flag of "OPT_A2STRVAL".
As the option value is not saved when the parameter is processed and is later
referenced when the option is printed a crash occurs.
The crash has been observed when using xl2tpd and the l2tp_ppp kernel module.

Additionally the arg argument was missing to the printer call in the print_option
utility when the option flag OPT_A2STRVAL is set.