ppp: fix pppol2tp option printing
PPPD crashes (SEGV) when the 'dump' or 'dryrun' options are specified and
an option defined internally as "o_special" with an option flag of
"OPT_A2STRVAL" is used.  The crash occurs because the option value is not
saved when the parameter is processed, but is then referenced when printed.
Additionally, the "arg" parameter is missing from the call to the "printer"
utility.  This was encountered using xl2tpd and the l2tp_ppp kernel module;
the option PPPD crashes on is "pppol2tp 8".

Modify pppd to correctly save the option value, and to call the printer
utility with the correct parameters.

Signed-off-by: Nathan Hintz <nlhintz@hotmail.com>

---

This patch was originally submitted to OpenWRT in:

	https://lists.openwrt.org/pipermail/openwrt-devel/2014-July/026794.html

and was recently accepted in the following changesets:

	https://dev.openwrt.org/changeset/45263
	https://dev.openwrt.org/changeset/45264

Note: This patch resolves issue #32 (https://github.com/paulusmack/ppp/issues/32)