linux: support non-standard baudrates
When an unsupported baudrate is specify, we can force it to the kernel
by using the TCSETS2 ioctl with the BOTHER cflag.

TCSETS2 and struct termios2 are available in asm/termbits.h and
asm/ioctls.h but including these headers conflict with termios.h which
is needed by sys-linux.c.
For that reason, we put the force_baudrate function in a dedicated file.

Signed-off-by: Sylvain Chouleur <sylvain.chouleur@gmail.com>