config: pass prefix unexpanded when generating build scripts
Rename ${DESTDIR} to ${prefix} and only let make expand it. This is
basically analogous to what autotools do. It will be useful for generating a
proper pkg-config file (where the prefix is called ${prefix} and is expected to
be unexpanded).

DESTDIR name was used inconsistently with what the rest of the world
does (where it modifies the installation prefix, sort of what our
${INSTROOT} does).

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>