pppd: Fix IPv6 default route code for Solaris
Commit 388597ee ("pppd: Add defaultroute6 and related options") added
code to pppd/sys-solaris.c which only works on Linux.  Solaris doesn't
allow the use of the SICORT* family of ioctls for IPv6.  They're legacy
IPv4 only.  Routing sockets are much more flexible than the ioctls.
This rewrites the Solaris code to use a routing socket to set the
default route.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>