Convert to ANSI C
This gets rid of the __P and __V macros that were used so that the
code was in theory compilable by a K&R C compiler, and converts the
function definitions to ANSI C style.  In fact there were already
quite a few function definitions in the ANSI C style, so it would not
have been compilable by a K&R C compiler in fact.

The Solaris and BSD kernel code modules have had __P removed but the
function definitions have not been converted.

There are some other minor changes here to remove warnings.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>