pppd: Use a run-time test to detect libutil availability on Linux
This tests whether logwtmp is declared in <utmp.h>, and if not we
assume we need to include our own logwtmp implementation.  We assume
that if logwtmp is provided by the system then we need -lutil to get
it (as is the case for glibc).

This fixes compilation with musl libc and other libcs that don't
provide logwtmp.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>