ipv6cp: Fix clearing previous IPv6 LL address in demand mode
Previous old IPv6 LL address is stored in the 'wo' structure. In 'go'/'ho'
is stored new address which needs to be assigned on interface.

Without this patch pppd in demand IPv6 mode always let previous IPv6 LL
address on interface and failed on error:

cif6addr: ioctl(SIOCDIFADDR): No such address

Because it tried to remove new address from interface which obviously was
not set.

Signed-off-by: Pali Rohár <pali@kernel.org>