EAP-TLS: Verify Subject or CommonName by suffix
This feature matches closely what OpenVPN and the network-manager-openvpn plugin does for
certificate verification. It allows the end user to configure the certificate to be
matched by it's common name (entire string), it's subject name, or the suffix of a subject name
the latter is especially useful if you are trying to match against a random server in a RADIUS
pool. Lastly, it also allow you to turn off the certificate matching all together.

tls-verify-method can have the following parameter values:
- none
- subject   - The entire subject, e.g. /CN=some.server.org
- name      - The entie common name, e.g. some.server.org
- suffix    - The latter part of a name, e.g. servers.org

Secondly, it also introduces a new parameter 'tls-verify-key-usage' which permits checking
of the 'server' or 'client' side attributes of nsCertType and X509 extended key attributes.
For example, in client mode, it will verify that received certificate has the 'server' side
attributes enabled.

Signed-off-by: Eivind Naess <eivnaes@yahoo.com>